control-plane/requirements.txt
fastapi==0.115.8
pydantic==2.10.6
pydantic-settings==2.7.1
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
